refactor: split candidates.rs into candidates/ module with CandidateStrategy
- Convert monolithic candidates.rs (707 lines) into candidates/ directory
- Extract standard.rs (standard Viterbi + history candidates)
- Extract predictive.rs (bigram chaining for Copilot-like completions)
- Extract neural.rs (speculative decode candidates, #[cfg(feature = "neural")])
- Add strategy.rs with CandidateStrategy enum for pluggable dispatch
- Maintain backward-compatible public API via wrapper functions in mod.rs
- Move unit tests: chain-related to predictive::tests, integration to tests.rs
- No FFI changes required
