+++ engine/src/candidates/standard.rs
@@ -1,0 +1,142 @@
+use std::collections::HashSet;
+
+use tracing::{debug, debug_span};
+
+use crate::converter::convert_nbest;
+use crate::dict::connection::ConnectionMatrix;
+use crate::dict::{DictEntry, Dictionary, TrieDictionary};
+use crate::user_history::UserHistory;
+
+use super::{generate_punctuation_candidates, punctuation_alternatives, CandidateResponse};
+
+/// Generate candidates for normal (non-punctuation) input.
+pub(super) fn generate_normal_candidates(
+    dict: &TrieDictionary,
+    conn: Option<&ConnectionMatrix>,
+    history: Option<&UserHistory>,
+    reading: &str,
+    max_results: usize,
+) -> CandidateResponse {
+    let mut surfaces = Vec::new();
+    let mut seen = HashSet::new();
+
+    // 1. N-best Viterbi conversion (pure statistical, no history bias).
+    //    History influence is applied at the candidate level (kana promotion,
+    //    predictions, lookup reordering) rather than distorting Viterbi paths,
+    //    which avoids stale history overriding the statistically best results.
+    let nbest = 5usize;
+    let paths = convert_nbest(dict, conn, reading, nbest);
+
+    let mut nbest_paths = Vec::new();
+    for path in &paths {
+        let joined: String = path.iter().map(|s| s.surface.as_str()).collect();
+        if !joined.is_empty() && seen.insert(joined.clone()) {
+            surfaces.push(joined);
+        }
+        nbest_paths.push(path.clone());
+    }
+
+    // 2. Reading (kana) — position depends on history boost.
+    //    If the user previously selected the hiragana form, promote it to
+    //    position 0 (inline preview) so it becomes the default candidate.
+    //    The kana may already exist in N-best (via fallback nodes); if so, move it.
+    let now = crate::user_history::now_epoch();
+    let kana_boost = history.map_or(0, |h| h.unigram_boost(reading, reading, now));
+    let kana_existing_pos = surfaces.iter().position(|s| s == reading);
+    if kana_boost > 0 {
+        match kana_existing_pos {
+            Some(0) => {} // already at top
+            Some(pos) => {
+                surfaces.remove(pos);
+                surfaces.insert(0, reading.to_string());
+            }
+            None => {
+                seen.insert(reading.to_string());
+                surfaces.insert(0, reading.to_string());
+            }
+        }
+    } else if kana_existing_pos.is_none() {
+        seen.insert(reading.to_string());
+        surfaces.push(reading.to_string());
+    }
+
+    // 3. Predictions (ranked by history if available)
+    let fetch_limit = if history.is_some() {
+        max_results.max(200)
+    } else {
+        max_results
+    };
+    let mut ranked = dict.predict_ranked(reading, fetch_limit, 1000);
+    if let Some(h) = history {
+        ranked.sort_by(|(r_a, e_a), (r_b, e_b)| {
+            let boost_a = h.unigram_boost(r_a, &e_a.surface, now);
+            let boost_b = h.unigram_boost(r_b, &e_b.surface, now);
+            boost_b.cmp(&boost_a).then(e_a.cost.cmp(&e_b.cost))
+        });
+        ranked.truncate(max_results);
+    }
+    for (_, entry) in &ranked {
+        if !entry.surface.is_empty() && seen.insert(entry.surface.clone()) {
+            surfaces.push(entry.surface.clone());
+        }
+    }
+
+    // 4. Dictionary lookup
+    let lookup_entries: Vec<&DictEntry> = if let Some(h) = history {
+        if let Some(entries) = dict.lookup(reading) {
+            let reordered = h.reorder_candidates(reading, entries);
+            // We need owned entries; collect surfaces
+            for entry in &reordered {
+                if seen.insert(entry.surface.clone()) {
+                    surfaces.push(entry.surface.clone());
+                }
+            }
+            Vec::new() // already processed
+        } else {
+            Vec::new()
+        }
+    } else if let Some(entries) = dict.lookup(reading) {
+        entries.iter().collect()
+    } else {
+        Vec::new()
+    };
+
+    for entry in &lookup_entries {
+        if seen.insert(entry.surface.clone()) {
+            surfaces.push(entry.surface.clone());
+        }
+    }
+
+    CandidateResponse {
+        surfaces,
+        paths: nbest_paths,
+    }
+}
+
+/// Unified candidate generation: handles both punctuation and normal input.
+pub fn generate(
+    dict: &TrieDictionary,
+    conn: Option<&ConnectionMatrix>,
+    history: Option<&UserHistory>,
+    reading: &str,
+    max_results: usize,
+) -> CandidateResponse {
+    let _span = debug_span!("generate_candidates", reading, max_results).entered();
+    if reading.is_empty() {
+        return CandidateResponse {
+            surfaces: Vec::new(),
+            paths: Vec::new(),
+        };
+    }
+
+    let resp = if punctuation_alternatives(reading).is_some() {
+        generate_punctuation_candidates(dict, history, reading, max_results)
+    } else {
+        generate_normal_candidates(dict, conn, history, reading, max_results)
+    };
+    debug!(
+        surface_count = resp.surfaces.len(),
+        path_count = resp.paths.len()
+    );
+    resp
+}
